fix: sort results
